pyproject.toml
[tool.poetry]
name = "ieqtests"
version = "0.1.0"
description = ""
authors = ["Ilya Verbilo <verbiloilyaa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
drf-yasg = "^1.21.5"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
